DOP-3578: Move tests to new directories, add to ci
