+++ .github/workflows/compile.yml
@@ -26,3 +26,4 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - run: npm ci
       - run: npm run lint
+      - run: npm run test
