+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
   "extension": ["ts"],
-  "spec": ["tests/**.test.ts", "src/**.test.ts", "src/**/**.test.ts"],
+  "spec": ["tests/**/**/**.test.ts"],
   "require": "ts-node/register"
 }
