+++ tests/integration/test_data/state-1/bi-connector-v2.json
