+++ tests/integration/test_data/state-2/bi-connector-v2.json
